Move reducers, actions, and store folders into a 'state' folder
