+++ jsconfig.json
@@ -3,7 +3,6 @@
     "baseUrl": "./",
     "paths": {
       "*": ["src/*"],
-      "actions/*": ["src/actions/*"],
       "httpClient/*": ["src/httpClient/*"],
       "components/*": ["src/components/*"],
       "middleware/*": ["src/middleware/*"],
@@ -13,11 +12,10 @@
       "hooks/*": ["src/hooks/*"],
       "locales/*": ["src/locales/*"],
       "pages/*": ["src/pages/*"],
-      "reducers/*": ["src/reducers/*"],
+      "state/*": ["src/state/*"],
       "reusableTests": ["../cypress/support/reusableTests"],
       "selectors/*": ["src/selectors/*"],
       "services/*": ["src/services/*"],
-      "store/*": ["src/store/*"],
       "stubs": ["../cypress/stubs"],
       "styles/*": ["src/styles/*"],
       "utils/*": ["src/utils/*"],
