+++ server/client.js
@@ -5,7 +5,7 @@ import { AppContainer } from 'react-hot-loader';
 import { IntlProvider } from 'react-intl';
 import includes from 'lodash/includes';
 import { PersistGate } from 'redux-persist/lib/integration/react';
-import configureStore from 'store/configureStore';
+import configureStore from 'state/store/configureStore';
 import App from 'components/App';
 import locales from 'locales';
 import { SUPPORTED_LANGUAGES, DEFAULT_LANGUAGE } from 'constants/constants';
