+++ webpack/shared/resolve.js
@@ -3,7 +3,6 @@ const path = require('path');
 module.exports = {
   extensions: ['*', '.js', '.jsx', '.json'],
   alias: {
-    actions: path.resolve(__dirname, '../../src/actions'),
     httpClient: path.resolve(__dirname, '../../src/httpClient'),
     assets: path.resolve(__dirname, '../../src/assets'),
     components: path.resolve(__dirname, '../../src/components'),
@@ -13,9 +12,8 @@ module.exports = {
     services: path.resolve(__dirname, '../../src/services'),
     hooks: path.resolve(__dirname, '../../src/hooks'),
     locales: path.resolve(__dirname, '../../src/locales'),
-    reducers: path.resolve(__dirname, '../../src/reducers'),
+    state: path.resolve(__dirname, '../../src/state'),
     selectors: path.resolve(__dirname, '../../src/selectors'),
-    store: path.resolve(__dirname, '../../src/store'),
     styles: path.resolve(__dirname, '../../src/styles'),
     utils: path.resolve(__dirname, '../../src/utils'),
     fixtures: path.resolve(__dirname, '../../cypress/fixtures'),
